+++ src/stories/plan/PlanMemo.stories.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import PlanMemo from '@/components/plan/modal/PlanMemo';
+import PlanMemo from '@/components/modal/plan/PlanMemo';
 
 export default {
   title: 'Plan/PlanMemo',
